Fixed issues with not being able to enter SUDO password

+++ updates.sh
@@ -1,2 +1,11 @@
 #!/bin/bash
-sudo apt install lxpolkit policykit-1 -y
+FILE=~/.config/autostart/lxpolkit.desktop
+if ! test -f "$FILE"; then
+    sudo apt install lxpolkit policykit-1 -y
+    echo "[Desktop Entry]
+     Name=LXPolKit
+     Exec=/bin/lxpolkit
+     Terminal=false
+     Type=Application" | sudo tee ~/.config/autostart/lxpolkit.desktop
+fi
+
